Many Updates
Bug Fixes
Tile Tearing fixed across the whole project, even for future levels.
Fixed a bug where the player is on top of the TV and other objects.
Fixed a bug where the Sticky Notes were on top of the player.
Fixed a bug where "Player Interaction" wasn't working.
Increased interaction boxes so you can interact with sticky notes from all angles, and a bit further away.
Persephone is a Side Sleeper is now canon.
Music now Loops
Physics table constrained to moving only in the Y axis and only active when "Final Notes" are active
Fixed issue where "doorOpenSFX" played when going between living room and kitchen.
Added Persephone cuddling Cerb in the beginning.
Added Cerb to the bed after opening cutscene.
Once Flashlight is picked up, "pot" is disabled and "shards" and "potShatterEffect" are enabled so the player gets a sound jump scare and the pot is shattered now in the hallway.
Fixed Door collisions so you dont enter a room and appear already 5 feet inside. You should now appear in the doorway.
Fixed bug where last Creepy note was FinalNote_01 from the final category instead of the last Creepy note of the same name.
Replaced stickynotes with pixel ones.

New Bugs
Cerb plushie needs to appear sooner on the bed. Right now, it appears when the player anim is active. I would like to set Cerb active when the player is no longer sleeping in the bed.
-